requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
httpx>=0.24.0
pydantic>=1.10.0
python-dotenv>=1.0.0
pytest>=7.0.0
google-genai>=0.8.0
fpdf2>=2.8.0
pymongo>=4.8.0
